[#S05P21A506-52] chore: Entity 수정

See merge request s05-blockchain/S05P21A506!16

+++ backend/src/main/java/com/a506/mirinae/domain/funding/FundingReq.java
@@ -11,7 +11,7 @@ public class FundingReq {
     String title;
     Long categoryId;
     String content;
-    Long goal;
+    Double goal;
     String thumbnail;
     String image;
     LocalDateTime startDatetime;
